+++ client/src/components/Nuka-carousel/nukaCarousel.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+import Carousel from 'nuka-carousel';
+
+export default class NukaCarousel extends Component {
+  render () {
+    return (
+      <Carousel autoplay={true} autoplayInterval={1000} slideIndex={5} slidesToShow={4}  >
+        <img src="https://via.placeholder.com/400/ffffff/c0392b/&text=slide1"/>
+        <img src="https://via.placeholder.com/400/ffffff/c0392b/&text=slide2"/>
+        <img src="https://via.placeholder.com/400/ffffff/c0392b/&text=slide3"/>
+        <img src="https://via.placeholder.com/400/ffffff/c0392b/&text=slide4"/>
+        <img src="https://via.placeholder.com/400/ffffff/c0392b/&text=slide5"/>
+        <img src="https://via.placeholder.com/400/ffffff/c0392b/&text=slide6"/>
+      </Carousel>
+    );
+  }
+}
